Rename `ProxyRequest` to `ProxyGetRequest` and rename mod to `middleware`
