+++ http-server/src/lib.rs
@@ -35,8 +35,8 @@ mod server;
 /// Common builders for RPC responses.
 pub mod response;
 
-/// Common tower middlewares exposed for RPC interaction.
-pub mod middlewares;
+/// Common tower middleware exposed for RPC interaction.
+pub mod middleware;
 
 pub use jsonrpsee_core::server::access_control::{AccessControl, AccessControlBuilder};
 pub use jsonrpsee_core::server::rpc_module::RpcModule;
